+++ src/django_qstash/schedules/admin.py
@@ -9,7 +9,13 @@ from django_qstash.schedules.models import TaskSchedule
 
 @admin.register(TaskSchedule)
 class TaskScheduleAdmin(admin.ModelAdmin):
-    list_display = ["schedule_id", "task_name"]
+    list_display = [
+        "schedule_id",
+        "display_task_name",
+        "display_task_path",
+        "cron",
+        "is_active",
+    ]
     readonly_fields = [
         "schedule_id",
         "task_name",
@@ -64,3 +70,11 @@ class TaskScheduleAdmin(admin.ModelAdmin):
         if not obj.schedule_id:
             return "No schedule ID yet"
         return services.get_task_schedule_from_qstash(obj, as_dict=True)
+
+    @admin.display(description="Task Name", ordering="name")
+    def display_task_name(self, obj: TaskSchedule) -> str:
+        return obj.name or obj.task_name or obj.task
+
+    @admin.display(description="Task Path", ordering="task")
+    def display_task_path(self, obj: TaskSchedule) -> str:
+        return obj.task or obj.task_name